rename backendauthrotators to rotators
