+++ internal/controller/rotators/aws_common_test.go
@@ -1,4 +1,4 @@
-package backendauthrotators
+package rotators
 
 import (
 	"fmt"
