+++ internal/controller/rotators/aws_oidc_rotator.go
@@ -1,4 +1,4 @@
-package backendauthrotators
+package rotators
 
 import (
 	"context"
