+++ internal/controller/rotators/aws_oidc_rotator_test.go
@@ -1,4 +1,4 @@
-package backendauthrotators
+package rotators
 
 import (
 	"context"
