+++ internal/controller/rotators/common.go
@@ -1,4 +1,4 @@
-package backendauthrotators
+package rotators
 
 import (
 	"context"
