+++ internal/controller/rotators/common_test.go
@@ -1,4 +1,4 @@
-package backendauthrotators
+package rotators
 
 import (
 	"context"
